Switch from integers to doubles

+++ ch-4/drill/drill.cpp
@@ -3,8 +3,8 @@
 int main(int, char **) {
   std::cout << "> ";
 
-  int val1 = 0;
-  int val2 = 0;
+  double val1 = 0;
+  double val2 = 0;
   while (std::cin >> val1 >> val2) {
     if (val1 < val2) {
       std::cout << "Smaller is: " << val1 << "\tLarger is: " << val2 << '\n';
